chore: prepare for publishing
