+++ cli/index.ts
@@ -2,8 +2,8 @@ import fs from "node:fs/promises";
 import yargs from "yargs";
 import path from "node:path";
 import { hideBin } from "yargs/helpers";
-import * as importer from "~/import";
-import * as exporter from "~/export";
+import * as importer from "./importer";
+import * as exporter from "./exporter";
 
 import { DEFAULT_VERSION } from "~/entities/import";
 
@@ -19,16 +19,16 @@ const argv = yargs(hideBin(process.argv))
 		},
 		version: { type: "string", alias: "v", default: DEFAULT_VERSION },
 	})
-	.check((argv) => importer.supports(argv.bank))
+	.check((argv) => importer.get(argv.bank))
 	.parseSync();
 
 async function main() {
-	const pdf = await fs.readFile(path.resolve(__dirname, argv.in));
+	const pdf = await fs.readFile(path.resolve(__dirname, '..', argv.in));
 
 	const rows = await importer.run(argv.bank, argv.version, pdf);
 	const buffer = await exporter.run(rows, argv.out); 
 
-	await fs.writeFile(path.resolve(__dirname, argv.out), buffer);
+	await fs.writeFile(path.resolve(__dirname, '..', argv.out), buffer);
 }
 
 main().catch(console.error);
