+++ lib/entities/currency.ts
