+++ lib/entities/row.ts
